+++ src/card/cards/combat_constraints.rs
@@ -229,3 +229,5 @@ pub const SKIRK_PROSPECTOR: CardDefinitionId = CardDefinitionId(2028);
 pub const OVERLOAD: CardDefinitionId = CardDefinitionId(2029);
 pub const PROHIBIT: CardDefinitionId = CardDefinitionId(2030);
 pub const PYROKINESIS: CardDefinitionId = CardDefinitionId(2031);
+pub const GOBLIN_VANDAL: CardDefinitionId = CardDefinitionId(2032);
+pub const BARBARIAN_RING: CardDefinitionId = CardDefinitionId(2033);