+++ src/card/sets/y1997/mod.rs
@@ -2,5 +2,6 @@
 
 pub(super) mod tempest;
 pub(super) mod visions;
+pub(super) mod weatherlight;
 
 pub(super) use super::{CardRecord, PrintingRecord};